Undoing experiment with BigDecimal and reverting to r120.

+++ branches/repr-expr/invextension/src/com/moneydance/modules/features/invextension/ComponentReport.java
@@ -27,8 +27,6 @@
  */
 package com.moneydance.modules.features.invextension;
 
-import java.math.BigDecimal;
-
 /**
  * Base class for security reports and composites.
  *
@@ -65,9 +63,9 @@ public abstract class ComponentReport {
                 && retMap.values().size() > 0) {
             int i = 0;
             for (Integer dateInt : retMap.keySet()) {
-                BigDecimal value = retMap.get(dateInt);
+                long value = retMap.get(dateInt);
                 dateIntsArray[i] = dateInt;
-                annRetValuesArray[i] = value.intValue();
+                annRetValuesArray[i] = value;
                 i++;
             }
         } else {
@@ -111,7 +109,7 @@ public abstract class ComponentReport {
             int cd = DateUtils.getDaysBetween(mdMap.firstKey(), mdMap.lastKey());
 
             for (Integer thisDateInt : mdMap.keySet()) {
-                Long cf = mdMap.get(thisDateInt).longValue();
+                Long cf = mdMap.get(thisDateInt);
                 int dayBetw = DateUtils.getDaysBetween(mdMap.firstKey(), thisDateInt);
                 double wSubI = ((double)cd - (double)dayBetw) / (double)cd;
                 weightCF += Math.round(wSubI * cf);
